🔨 유틸 함수 추가 및 정리
